Move handle_exception
